Move UNIVERSE to betty.universe
