+++ betty/console/project.py
@@ -19,7 +19,7 @@ from betty.portable.file import assert_load_file
 from betty.project import Project
 from betty.project.data import ProjectConfiguration
 from betty.serde import SerializerDefinition
-from betty.service.level.universe import UNIVERSE
+from betty.universe import UNIVERSE
 from betty.user import User
