+++ betty/tests/plugins/loader/demo/test_project.py
@@ -6,8 +6,8 @@ from betty.copyright_notice import CopyrightNotice
 from betty.entity.collection.pool import EntityPool
 from betty.license import License
 from betty.plugins.loader.demo import LoadAncestry
-from betty.service.level.universe import UNIVERSE
 from betty.test_utils.job import do
+from betty.universe import UNIVERSE
 
 
 @pytest.mark.usefixtures("demo_project_aioresponses")
